+++ src/Tienda/Productos/TipoProductos/ProductoEnvasado.java
@@ -10,8 +10,8 @@ public class ProductoEnvasado extends Producto {
             String identificador,
             String descripcion,
             int stock,
-            float precioUnitario,
-            float porcentajeGanancia,
+            double precioUnitario,
+            double porcentajeGanancia,
             boolean disponible,
             String tipoEnvase,
             boolean importado)
